Module Log_eventlog

include Win_eventlog__Log_eventlog

val reporter : eventlog:Win_eventlog.Eventlog.t -> ?⁠category:int -> ?⁠event:int -> unit -> Logs.reporter

Construct a Logs reporter which sends logs to the Windows event log, optionally with the given category and event ids set.